elf/src/requirements.txt
fastapi
python-multipart
starlette
uvicorn
pillow
pytesseract
python-dotenv
scikit-learn
tensorflow==2.15.0
keras==2.15.0
numpy
llama-cpp-python